Type reducer
